+++ extensions/ql-vscode/src/legacy-query-server/run-queries.ts
@@ -1,5 +1,3 @@
-import { createHash } from "crypto";
-import { readFile } from "fs-extra";
 import * as tmp from "tmp-promise";
 import { basename, join } from "path";
 import { CancellationToken, Uri } from "vscode";
@@ -21,10 +19,7 @@ import * as messages from "../pure/legacy-messages";
 import { InitialQueryInfo, LocalQueryInfo } from "../query-results";
 import * as qsClient from "./queryserver-client";
 import { getErrorMessage } from "../pure/helpers-pure";
-import {
-  compileDatabaseUpgradeSequence,
-  upgradeDatabaseExplicit,
-} from "./upgrades";
+import { compileDatabaseUpgradeSequence } from "./upgrades";
 import { QueryEvaluationInfo, QueryWithResults } from "../run-queries-shared";
 
 /**
@@ -233,57 +228,6 @@ export async function clearCacheInDatabase(
   return qs.sendRequest(messages.clearCache, params, token, progress);
 }
 
-/**
- * Compare the dbscheme implied by the query `query` and that of the current database.
- * - If they are compatible, do nothing.
- * - If they are incompatible but the database can be upgraded, suggest that upgrade.
- * - If they are incompatible and the database cannot be upgraded, throw an error.
- */
-async function checkDbschemeCompatibility(
-  cliServer: cli.CodeQLCliServer,
-  qs: qsClient.QueryServerClient,
-  query: QueryInProgress,
-  qlProgram: messages.QlProgram,
-  dbItem: DatabaseItem,
-  progress: ProgressCallback,
-  token: CancellationToken,
-): Promise<void> {
-  const searchPath = getOnDiskWorkspaceFolders();
-
-  if (dbItem.contents?.dbSchemeUri !== undefined) {
-    const { finalDbscheme } = await cliServer.resolveUpgrades(
-      dbItem.contents.dbSchemeUri.fsPath,
-      searchPath,
-      false,
-    );
-    const hash = async function (filename: string): Promise<string> {
-      return createHash("sha256")
-        .update(await readFile(filename))
-        .digest("hex");
-    };
-
-    // At this point, we have learned about three dbschemes:
-
-    // the dbscheme of the actual database we're querying.
-    const dbschemeOfDb = await hash(dbItem.contents.dbSchemeUri.fsPath);
-
-    // the dbscheme of the query we're running, including the library we've resolved it to use.
-    const dbschemeOfLib = await hash(query.queryDbscheme);
-
-    // the database we're able to upgrade to
-    const upgradableTo = await hash(finalDbscheme);
-
-    if (upgradableTo != dbschemeOfLib) {
-      reportNoUpgradePath(qlProgram, query);
-    }
-
-    if (upgradableTo == dbschemeOfLib && dbschemeOfDb != dbschemeOfLib) {
-      // Try to upgrade the database
-      await upgradeDatabaseExplicit(qs, dbItem, progress, token);
-    }
-  }
-}
-
 function reportNoUpgradePath(
   qlProgram: messages.QlProgram,
   query: QueryInProgress,
@@ -309,11 +253,8 @@ async function compileNonDestructiveUpgrade(
     throw new Error("Database is invalid, and cannot be upgraded.");
   }
 
-  // When packaging is used, dependencies may exist outside of the workspace and they are always on the resolved search path.
-  // When packaging is not used, all dependencies are in the workspace.
-  const upgradesPath = (await qs.cliServer.cliConstraints.supportsPackaging())
-    ? qlProgram.libraryPath
-    : getOnDiskWorkspaceFolders();
+  // Dependencies may exist outside of the workspace and they are always on the resolved search path.
+  const upgradesPath = qlProgram.libraryPath;
 
   const { scripts, matchesTarget } = await qs.cliServer.resolveUpgrades(
     dbItem.contents.dbSchemeUri.fsPath,
@@ -409,33 +350,27 @@ export async function compileAndRunQueryAgainstDatabase(
   const metadata = await tryGetQueryMetadata(cliServer, qlProgram.queryPath);
 
   let availableMlModels: cli.MlModelInfo[] = [];
-  if (!(await cliServer.cliConstraints.supportsResolveMlModels())) {
-    void extLogger.log(
-      "Resolving ML models is unsupported by this version of the CLI. Running the query without any ML models.",
-    );
-  } else {
-    try {
-      availableMlModels = (
-        await cliServer.resolveMlModels(
-          diskWorkspaceFolders,
-          initialInfo.queryPath,
-        )
-      ).models;
-      if (availableMlModels.length) {
-        void extLogger.log(
-          `Found available ML models at the following paths: ${availableMlModels
-            .map((x) => `'${x.path}'`)
-            .join(", ")}.`,
-        );
-      } else {
-        void extLogger.log("Did not find any available ML models.");
-      }
-    } catch (e) {
-      const message =
-        `Couldn't resolve available ML models for ${qlProgram.queryPath}. Running the ` +
-        `query without any ML models: ${e}.`;
-      void showAndLogErrorMessage(message);
+  try {
+    availableMlModels = (
+      await cliServer.resolveMlModels(
+        diskWorkspaceFolders,
+        initialInfo.queryPath,
+      )
+    ).models;
+    if (availableMlModels.length) {
+      void extLogger.log(
+        `Found available ML models at the following paths: ${availableMlModels
+          .map((x) => `'${x.path}'`)
+          .join(", ")}.`,
+      );
+    } else {
+      void extLogger.log("Did not find any available ML models.");
     }
+  } catch (e) {
+    const message =
+      `Couldn't resolve available ML models for ${qlProgram.queryPath}. Running the ` +
+      `query without any ML models: ${e}.`;
+    void showAndLogErrorMessage(message);
   }
 
   const hasMetadataFile = await dbItem.hasMetadataFile();
@@ -452,29 +387,16 @@ export async function compileAndRunQueryAgainstDatabase(
 
   let upgradeDir: tmp.DirectoryResult | undefined;
   try {
-    let upgradeQlo;
-    if (await cliServer.cliConstraints.supportsNonDestructiveUpgrades()) {
-      upgradeDir = await tmp.dir({ dir: upgradesTmpDir, unsafeCleanup: true });
-      upgradeQlo = await compileNonDestructiveUpgrade(
-        qs,
-        upgradeDir,
-        query,
-        qlProgram,
-        dbItem,
-        progress,
-        token,
-      );
-    } else {
-      await checkDbschemeCompatibility(
-        cliServer,
-        qs,
-        query,
-        qlProgram,
-        dbItem,
-        progress,
-        token,
-      );
-    }
+    upgradeDir = await tmp.dir({ dir: upgradesTmpDir, unsafeCleanup: true });
+    const upgradeQlo = await compileNonDestructiveUpgrade(
+      qs,
+      upgradeDir,
+      query,
+      qlProgram,
+      dbItem,
+      progress,
+      token,
+    );
     let errors;
     try {
       errors = await query.compile(qs, qlProgram, progress, token);